+++ src/index.ts
@@ -2,4 +2,4 @@ export * from './wrap'
 export * from './config'
 export * from './matchers'
 export * from './models'
-export * from './mswExtension'
+export * from './network/mockNetwork'
