+++ src/network/printRequest.ts
@@ -1,0 +1,15 @@
+import { WrapRequest } from '../models'
+import chalk from 'chalk'
+
+const printRequest = (request: WrapRequest) => {
+  return console.warn(`
+${chalk.white.bold.bgRed('wrapito')} ${chalk.redBright.bold(
+    'cannot find any mock matching:',
+  )}
+  ${chalk.greenBright(`URL: ${request.url}`)}
+  ${chalk.greenBright(`METHOD: ${request.method?.toLowerCase()}`)}
+  ${chalk.greenBright(`REQUEST BODY: ${request._bodyInit}`)}
+ `)
+}
+
+export { printRequest }
