+++ src/wrap.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 import { getConfig } from './config'
 import { getOptions, updateOptions } from './options'
 import type { Extension, Extensions, Response, Wrap, WrapExtensionAPI } from './models'
-import { createMswNetworkMocker } from './mswExtension'
+import { createMswNetworkMocker } from './network/mockNetwork'
 
 
 const wrap = (component: unknown): Wrap => {
